+++ packages/react-icons-file-type/bundle-size/file-type-icon-headless.fixture.js
@@ -1,5 +1,0 @@
-import { FileTypeIcon } from '@fluentui/react-icons-file-type/headless';
-
-console.log(FileTypeIcon);
-
-export default { name: 'FileTypeIcon (headless, no Griffel)' };